Random musings from cynical old LEETH
I'm going to try and recall how/what we did things to point out why it was a good idea at the time. This brainꟻart will be all moved to the legacy section over time as I write as stream of consciousness which is not pontification from on-high truth beyond repute. I leave that to the obviously smarter, faster (but debatably cuter) NeoClass. I call it this because every new technology destroys old jobs, and creates new roles. Nobody knew what SEO was b4 google. So the new classes (and titles) are yours to homestead after the pioneers have their backs shot with indian arrows. This will be my 3rd and final hackathon.